fix: substitui double por BigDecimal no gerarExtratoFinanceiro e RelatorioFinanceiroCompletoDTO

+++ src/main/java/agendamentoDeClienteBarbearia/dtos/RelatorioFinanceiroCompletoDTO.java
@@ -4,12 +4,13 @@ package agendamentoDeClienteBarbearia.dtos;
 
 import agendamentoDeClienteBarbearia.dtosResponse.DetalhamentoAgendamentoDTO;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record RelatorioFinanceiroCompletoDTO(
-        Double totalFaturamento,
-        Double totalCasa,
-        Double totalComissoes,
+        BigDecimal totalFaturamento,
+        BigDecimal totalCasa,
+        BigDecimal totalComissoes,
         int quantidadeAtendimentos,
         List<DetalhamentoAgendamentoDTO> extrato
 ) {}
